- fix settings page local name not being updated - add user post

+++ src/app/(logged)/component.client.tsx
@@ -1,35 +1,29 @@
 'use client';
-import { addGlobalMessageSchema, type addGlobalMessageValues } from '@/lib/validation';
+import { addGlobalMessageSchema, type AddGlobalMessageValues } from '@/lib/validation';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Form, FormControl, FormField, FormItem, FormMessage } from '@/components/ui/form';
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '@/components/ui/dropdown-menu';
 import { Button } from '@/components/ui/button';
 import { Textarea } from '@/components/ui/textarea';
 import { useForm } from 'react-hook-form';
-import { addGlobalMessage, deleteGlobalMessage } from './actions';
-import { useEffect } from 'react';
+import { addGlobalMessage, deleteGlobalMessage, reportGlobalMessage } from './actions';
 import { Ellipsis, Flag, Trash2 } from 'lucide-react';
 import { MessageType } from './page';
+import { toast } from 'sonner';
 
-export function Action({ action }: { action: () => any }) {
- useEffect(() => {
-  action();
- }, [action]);
- return null;
-}
-
-export default function ClientComponent() {
- const form = useForm<addGlobalMessageValues>({
+export default function Client() {
+ const form = useForm<AddGlobalMessageValues>({
   resolver: zodResolver(addGlobalMessageSchema),
   defaultValues: { content: '' },
  });
 
- async function onSubmit(data: addGlobalMessageValues) {
+ async function onSubmit(data: AddGlobalMessageValues) {
   try {
    await addGlobalMessage(data);
   } catch (err) {
    console.log(err);
   }
+  form.reset();
  }
  return (
   <Form {...form}>
@@ -69,18 +63,32 @@ export function MessageInteractionDropdown(props: MessageInteractionDropdownProp
      <DropdownMenuItem
       className='flex flex-row items-center gap-4'
       onClick={() => {
-       /* if (confirm('Are you sure you want to delete this message?'))  */
-       deleteGlobalMessage(props.message.id, props.message.authorId);
+       if (confirm('Are you sure you want to delete this message?')) {
+        deleteGlobalMessage(props.message.id, props.message.authorId);
+        toast.success('Your message was deleted successfully!');
+       }
       }}
      >
       <Trash2 width={20} />
       <span>Delete</span>
      </DropdownMenuItem>
     )}
-    <DropdownMenuItem className='flex flex-row items-center gap-4'>
-     <Flag width={20} />
-     <span>Report</span>
-    </DropdownMenuItem>
+    {!props.isMessageAuthor && (
+     <DropdownMenuItem
+      className='flex flex-row items-center gap-4'
+      onClick={async () => {
+       if (confirm('Are you sure you want to report this message?')) {
+        const reportSuccess = await reportGlobalMessage(props.message.id, props.message.authorId);
+
+        if (reportSuccess) toast.success('Message reported!');
+        else toast.error('Already reported this message!');
+       }
+      }}
+     >
+      <Flag width={20} />
+      <span>Report</span>
+     </DropdownMenuItem>
+    )}
    </DropdownMenuContent>
   </DropdownMenu>
  );
